requirements.txt
# requirements.txt

fastapi
sqlalchemy
uvicorn[standard]
fastapi-jwt-auth
passlib
python-jose
loguru
pydantic
email-validator
bcrypt
slowapi

